\name{MovieSpider-package}
\alias{MovieSpider-package}
\alias{MovieSpider}
\docType{package}
\title{
\packageTitle{MovieSpider}
}
\description{
\packageDescription{MovieSpider}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MovieSpider}
\packageIndices{MovieSpider}
}
\author{
\packageAuthor{MovieSpider}

Maintainer: \packageMaintainer{MovieSpider}
}
